Fix access to activity class in onAttachedToActivity & add result to response
